+++ nemo_retriever/tests/__init__.py
@@ -19,6 +19,7 @@ __all__ = [
     "_have_ffmpeg_binary",
     "is_ffmpeg_png_encoder_available",
     "_have_ffmpeg_binary_for_png_frames",
+    "_make_test_mp4_with_av",
 ]
 
 
@@ -76,3 +77,30 @@ def is_ffmpeg_png_encoder_available() -> bool:
 def _have_ffmpeg_binary_for_png_frames() -> bool:
     """For pytest skips on paths that call ``MediaInterface.extract_frames`` (PNG output)."""
     return is_media_extract_available() and is_ffmpeg_png_encoder_available()
+
+
+def _make_test_mp4_with_av(path: Path, duration_sec: int = 5) -> None:
+    """Synthetic MP4 with video+audio; ``mpeg4`` avoids requiring ``libx264``."""
+    cmd = [
+        "ffmpeg",
+        "-y",
+        "-loglevel",
+        "error",
+        "-f",
+        "lavfi",
+        "-i",
+        f"testsrc=duration={duration_sec}:size=320x240:rate=30",
+        "-f",
+        "lavfi",
+        "-i",
+        f"sine=frequency=440:duration={duration_sec}",
+        "-c:v",
+        "mpeg4",
+        "-q:v",
+        "5",
+        "-c:a",
+        "aac",
+        "-shortest",
+        str(path),
+    ]
+    subprocess.run(cmd, check=True)
