+++ nemo_retriever/tests/test_video_pipeline_batch.py
@@ -6,7 +6,6 @@
 
 from __future__ import annotations
 
-import subprocess
 from pathlib import Path
 from unittest.mock import MagicMock, patch
 
@@ -14,6 +13,7 @@ import pandas as pd
 import pytest
 
 from tests import _have_ffmpeg_binary_for_png_frames
+from tests import _make_test_mp4_with_av
 from nemo_retriever.params import (
     ASRParams,
     AudioChunkParams,
@@ -56,34 +56,6 @@ def test_run_video_pipeline_forces_audio_demux_chunk_params_without_ffmpeg() -> 
     assert chunk_params.split_type == "time"
     assert chunk_params.split_interval == 10
     assert not out.empty
-
-
-def _make_test_mp4_with_av(path: Path, duration_sec: int = 5) -> None:
-    """Synthetic MP4 with video+audio; ``mpeg4`` avoids requiring ``libx264``."""
-    cmd = [
-        "ffmpeg",
-        "-y",
-        "-loglevel",
-        "error",
-        "-f",
-        "lavfi",
-        "-i",
-        f"testsrc=duration={duration_sec}:size=320x240:rate=30",
-        "-f",
-        "lavfi",
-        "-i",
-        f"sine=frequency=440:duration={duration_sec}",
-        "-c:v",
-        "mpeg4",
-        "-q:v",
-        "5",
-        "-c:a",
-        "aac",
-        "-shortest",
-        str(path),
-    ]
-    subprocess.run(cmd, check=True)
-
 
 @pytest.mark.skipif(
     not _have_ffmpeg_binary_for_png_frames(),
